\name{wythamewa-package}
\alias{wythamewa-package}
\alias{wythamewa} 
\docType{package}
\title{Wytham Woods social learning package}
\description{
  .
}

\details{
  \tabular{ll}{
    Package: \tab wythamewa\cr
    Type: \tab Package\cr
    Version: \tab 1\cr
    Date: \tab 10 July 2015\cr
    License: \tab GPL-3 \cr
  }
  
}
\author{
  Richard McElreath
}
\references{
  .
}

\examples{}  

\seealso{
} 
\keyword{wythamewa} 
\keyword{package} 