\name{WythamUnequal}
\alias{WythamUnequal}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Wytham Woods Parus major behavior data}
\description{
  .
}
\usage{
data(WythamUnequal)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
    \enumerate{
        \item herd : identifier for individual herd
        \item stock : categorical small or large stock
        \item n : number of animals at beginning of observation period
        \item mortality : number of animals that died in observation period
    }
}
\value{
}
\references{}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

