\name{WythamUnequal}
\alias{WythamUnequal}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Wytham Woods Parus major behavior data}
\description{
  .
}
\usage{
data(WythamUnequal)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
    \enumerate{
        \item N : Number of individual solves
        \item N_birds : Number of unique birds
        \item bird : Unique identifier for each bird
        \item solve_hi : 0/1 indicator of solving high-payoff option
        \item s_prop : Proportion of high-payoff solves observed. -1 indicates no recent social cues at time of solve.
        \item A_init : Initial attraction scores derived from prior experiments with same individuals.
        \item age : Each bird's age (standardized)
    }
}
\value{
}
\references{}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

